+++ apps/pay/src/utils/plugins/i18n/helpers/index.ts
@@ -4,6 +4,11 @@ import i18n from "@pay/utils/plugins/i18n";
 
 export async function loadLocaleMessages(locale: string) {
 	try {
+		if (!i18n.global.availableLocales.includes("en")) {
+			const fallbackMessages = await import(`@pay/utils/plugins/i18n/locales/en.json`);
+			i18n.global.setLocaleMessage("en", fallbackMessages.default);
+		}
+
 		const messages = await import(`@pay/utils/plugins/i18n/locales/${locale}.json`);
 		i18n.global.setLocaleMessage(locale, messages.default);
 		const elementHtml = document.querySelector("html");